Move file lock code to separate file
