+++ pkg/util/fileutil/fileutil.go
@@ -1,11 +1,7 @@
 package fileutil
 
 import (
-	"bytes"
-	"fmt"
-	"io"
 	"os"
-	"syscall"
 )
 
 // File exists has three different return cases that should be handled:
@@ -27,94 +23,4 @@ func FileExists(filename string) (bool, error) {
 	}
 
 	return !info.IsDir(), nil
-}
-
-// WriteLockedFD uses the flock() syscall to safely write to an open file as
-// long as other users of the file are also using flock()-based access.
-//
-// WriteLocked will block until it gets lock access.
-//
-// For the reasons outlined best in https://lwn.net/Articles/586904/ this uses
-// flock() instead of fcntl(). The ability to lock byte ranges is not necessary
-// and flock() has better behavior.
-func WriteLockedFD(f *os.File, data []byte) (int, error) {
-	// For the reasons outlined best in https://lwn.net/Articles/586904/ we're
-	// going to use flock() instead of fcntl() because we want a whole-file lock.
-	if err := syscall.Flock(int(f.Fd()), syscall.LOCK_EX); err != nil {
-		return 0, fmt.Errorf("unexpected error flock()-ing with EX: %w", err)
-	}
-
-	n, err := f.Write(data)
-	if err != nil {
-		return n, fmt.Errorf("writing data: %w", err)
-	}
-
-	if err := syscall.Flock(int(f.Fd()), syscall.LOCK_UN); err != nil {
-		return n, fmt.Errorf("unexpected error flock()-ing with UN: %w", err)
-	}
-
-	return n, nil
-}
-
-// WriteLocked uses the flock() syscall to safely write to a file as long as
-// other users of the file are also using flock()-based access.
-//
-// WriteLocked will block until it gets lock access.
-//
-// For the reasons outlined best in https://lwn.net/Articles/586904/ this uses
-// flock() instead of fcntl(). The ability to lock byte ranges is not necessary
-// and flock() has better behavior.
-func WriteLocked(filename string, data []byte) (int, error) {
-	file, err := os.OpenFile(filename, os.O_RDWR|os.O_CREATE, 0600)
-	if err != nil {
-		return 0, fmt.Errorf("opening %s: %w", filename, err)
-	}
-	defer file.Close()
-
-	return WriteLockedFD(file, data)
-}
-
-// ReadLockedFD uses the flock() syscall to safely read from an open file as
-// long as other users of the file are also using flock()-based access.
-//
-// ReadLocked will block until it gets lock access.
-//
-// For the reasons outlined best in https://lwn.net/Articles/586904/ this uses
-// flock() instead of fcntl(). The ability to lock byte ranges is not necessary
-// and flock() has better behavior.
-func ReadLockedFD(f *os.File) ([]byte, error) {
-	// For the reasons outlined best in https://lwn.net/Articles/586904/ we're
-	// going to use flock() instead of fcntl() because we want a whole-file lock.
-	if err := syscall.Flock(int(f.Fd()), syscall.LOCK_SH); err != nil {
-		return nil, fmt.Errorf("unexpected error flock()-ing with SH: %w", err)
-	}
-
-	buf := bytes.NewBuffer(nil)
-	if _, err := io.Copy(buf, f); err != nil {
-		return nil, fmt.Errorf("copying data out of file: %w", err)
-	}
-
-	if err := syscall.Flock(int(f.Fd()), syscall.LOCK_UN); err != nil {
-		return nil, fmt.Errorf("unexpected error flock()-ing with UN: %w", err)
-	}
-
-	return buf.Bytes(), nil
-}
-
-// ReadLocked uses the flock() syscall to safely read from a file as long as
-// other users of the file are also using flock()-based access.
-//
-// ReadLocked will block until it gets lock access.
-//
-// For the reasons outlined best in https://lwn.net/Articles/586904/ this uses
-// flock() instead of fcntl(). The ability to lock byte ranges is not necessary
-// and flock() has better behavior.
-func ReadLocked(filename string) ([]byte, error) {
-	file, err := os.OpenFile(filename, os.O_RDONLY, 0600)
-	if err != nil {
-		return nil, fmt.Errorf("opening %s: %w", filename, err)
-	}
-	defer file.Close()
-
-	return ReadLockedFD(file)
 }
